+++ source/_internals/createPath.js
@@ -1,3 +1,7 @@
+import { isInteger } from './isInteger.js'
+
 export function createPath(path, delimiter = '.'){
-  return typeof path === 'string' ? path.split(delimiter) : path
+  return typeof path === 'string' ?
+    path.split(delimiter).map(x => isInteger(x) ? Number(x) : x) :
+    path
 }