+++ source/_internals/isInteger.js
@@ -3,3 +3,8 @@ function _isInteger(n){
 }
 
 export const isInteger = Number.isInteger || _isInteger
+
+/**
+ * Check if `index` is integer even if it is a string.
+ */
+export const isIndexInteger = index => Number.isInteger(Number(index))