+++ source/addIndex-spec.ts
@@ -1,5 +1,4 @@
-import * as R from 'rambda'
-import * as Ramda from 'ramda'
+import {addIndex, map, pipe} from 'rambda'
 
 describe('R.addIndex', () => {
   it('happy', () => {
@@ -11,23 +10,15 @@ describe('R.addIndex', () => {
 
       return willReturn
     }
-    const mapIndexed = R.addIndex(mapFn)
+    const mapIndexed = addIndex(mapFn)
     const fn = (val: number, idx: number, list: number[]) =>
       val + idx + 5 + list[0]
     const result = mapIndexed(fn)([1, 2, 3])
     result // $ExpectType any[]
   })
-  it('with Ramda.pipe', () => {
-    const result = Ramda.pipe(
-      Ramda.addIndex(R.map)((x: number, i: number) => {
-        return x + i
-      })
-    )([1, 2, 3])
-    result // $ExpectType unknown
-  })
-  it('with Rambda.pipe', () => {
-    const result = R.pipe(
-      R.addIndex(Ramda.map)((x: number, i: number) => {
+  it('with pipe', () => {
+    const result = pipe(
+      addIndex(map)((x: number, i: number) => {
         return x + i
       })
     )([1, 2, 3])